# improve cache performance for single date queries