--- a/lib/holidays/definition/repository/cache.rb
+++ b/lib/holidays/definition/repository/cache.rb
@@ -10,15 +10,21 @@ module Holidays
           raise ArgumentError unless cache_data
 
           @cache_range[options] = start_date..end_date
-          @cache[options] = cache_data
+          @cache[options] = cache_data.each_with_object({}) do |holiday, cache|
+            date = holiday[:date]
+            holidays = cache.fetch(:holidays, [])
+            cache[date] = holidays + [holiday]
+          end
         end
 
         def find(start_date, end_date, options)
-          if range = @cache_range[options]
-            if range.begin <= start_date && range.end >= end_date
-              return @cache[options].select do |holiday|
-                holiday[:date] >= start_date && holiday[:date] <= end_date
-              end
+          if in_cache_range?(start_date, end_date, options)
+            if start_date == end_date
+              @cache[options].fetch(start_date, [])
+            else
+              @cache[options].select do |date, holidays|
+                date >= start_date && date <= end_date
+              end.flat_map { |date, holidays| holidays }
             end
           end
         end
@@ -26,6 +32,17 @@ module Holidays
         def reset!
           @cache = {}
           @cache_range = {}
+        end
+
+        private
+
+        def in_cache_range?(start_date, end_date, options)
+          range = @cache_range[options]
+          if range
+            range.begin <= start_date && range.end >= end_date
+          else
+            false
+          end
         end
       end
     end