--- a/test/holidays/definition/repository/test_cache.rb
+++ b/test/holidays/definition/repository/test_cache.rb
@@ -7,20 +7,55 @@ class CacheRepoTests < Test::Unit::TestCase
     @subject = Holidays::Definition::Repository::Cache.new
   end
 
+  def test_find_supports_overlapping_holidays
+    start_date = Date.civil(2015, 1, 1)
+    end_date = Date.civil(2015, 7, 1)
+    cache_data = [
+      {:date=>Date.civil(2015, 1, 1), :name=>"New Year's Day A", :regions=>[:us]},
+      {:date=>Date.civil(2015, 1, 1), :name=>"New Year's Day B", :regions=>[:us]}
+    ]
+    options = :us
+
+    @subject.cache_between(start_date, end_date, cache_data, options)
+
+    assert_equal(cache_data, @subject.find(start_date, end_date, options))
+  end
+
+  def test_find_returns_correct_cache_data_for_a_period
+    start_date = Date.civil(2015, 1, 1)
+    end_date = Date.civil(2015, 7, 1)
+    cache_data = [{:date=>Date.civil(2015, 1, 1), :name=>"New Year's Day", :regions=>[:us]}]
+    options = :us
+
+    @subject.cache_between(start_date, end_date, cache_data, options)
+
+    assert_equal(cache_data, @subject.find(start_date, end_date, options))
+  end
+
   def test_find_returns_correct_cache_data
     start_date = Date.civil(2015, 1, 1)
     end_date = Date.civil(2015, 1, 1)
-    cache_data =  [{:date=>Date.civil(2015, 1, 1), :name=>"New Year's Day", :regions=>[:us]}]
+    cache_data = [{:date=>Date.civil(2015, 1, 1), :name=>"New Year's Day", :regions=>[:us]}]
     options = :us
     @subject.cache_between(start_date, end_date, cache_data, options)
 
     assert_equal(cache_data, @subject.find(start_date, end_date, options))
   end
 
+  def test_find_returns_nil_if_no_match_is_found_for_a_period
+    start_date = Date.civil(2015, 1, 1)
+    end_date = Date.civil(2015, 7, 1)
+    cache_data = [{:date=>Date.civil(2015, 1, 1), :name=>"New Year's Day", :regions=>[:us]}]
+    options = :us
+    @subject.cache_between(start_date, end_date, cache_data, options)
+
+    assert_nil(@subject.find(Date.civil(2015, 7, 1), Date.civil(2015, 12, 1), options))
+  end
+
   def test_find_returns_nil_if_no_match_is_found
     start_date = Date.civil(2015, 1, 1)
     end_date = Date.civil(2015, 1, 1)
-    cache_data =  [{:date=>Date.civil(2015, 1, 1), :name=>"New Year's Day", :regions=>[:us]}]
+    cache_data = [{:date=>Date.civil(2015, 1, 1), :name=>"New Year's Day", :regions=>[:us]}]
     options = :us
     @subject.cache_between(start_date, end_date, cache_data, options)
 
@@ -30,7 +65,7 @@ class CacheRepoTests < Test::Unit::TestCase
   def test_cache_between_returns_error_if_dates_are_missing
     start_date = Date.civil(2015, 1, 1)
     end_date = Date.civil(2015, 1, 1)
-    cache_data =  [{:date=>Date.civil(2015, 1, 1), :name=>"New Year's Day", :regions=>[:us]}]
+    cache_data = [{:date=>Date.civil(2015, 1, 1), :name=>"New Year's Day", :regions=>[:us]}]
     options = :us
 
     assert_raise ArgumentError do
@@ -45,7 +80,7 @@ class CacheRepoTests < Test::Unit::TestCase
   def test_cache_between_returns_error_if_dates_are_invalid
     start_date = Date.civil(2015, 1, 1)
     end_date = Date.civil(2015, 1, 1)
-    cache_data =  [{:date=>Date.civil(2015, 1, 1), :name=>"New Year's Day", :regions=>[:us]}]
+    cache_data = [{:date=>Date.civil(2015, 1, 1), :name=>"New Year's Day", :regions=>[:us]}]
     options = :us
 
     assert_raise ArgumentError do
@@ -70,7 +105,7 @@ class CacheRepoTests < Test::Unit::TestCase
   def test_reset_clears_cache
     start_date = Date.civil(2015, 1, 1)
     end_date = Date.civil(2015, 1, 1)
-    cache_data =  [{:date=>Date.civil(2015, 1, 1), :name=>"New Year's Day", :regions=>[:us]}]
+    cache_data = [{:date=>Date.civil(2015, 1, 1), :name=>"New Year's Day", :regions=>[:us]}]
     options = :us
     @subject.cache_between(start_date, end_date, cache_data, options)