The Holidays gem performs a lot of single date queries, these are done by using the date extensions and calling `Date#holiday?` or by calling `Holidays.on?`.

The single date queries delegate to `Holidays.between(date, date)`, which in turn uses the underlying cache (if available).

The majority of use cases I have seen for the holiday gem in my organization were single date queries, which are very inefficient with the given cache structure.

The cache is currently an array of holidays, its lookup method has a `O(n)` time complexity where `n` is the amount of holidays currently in the cache.
As one would expect the performance impact is linear to the size of the cache.

This simple optimization changes the cache's data store from an array to a lookup table.

## Benchmarks
```ruby
date = Date.new(2017, 1, 1)
Holidays.cache_between(
  Date.new(2017, 1, 1), 
  Date.new(2017 + years , 1, 1), 
  :us, 
  :observed
)
Benchmark.measure { 10000.times.each { date.holiday?(:us, :observed) } }.real
```

As a baseline, this computation takes **4.280** seconds without any cache.

### Before

| Years  | Cache Size | Time (seconds) |
| ------ | ----------- | ---------------- |
| 1 | 11 | 0.069 |
| 5 | 51 | 0.168 |
| 10 | 101 | 0.292 |
| 15 | 151 |  0.437 |
| 20 | 201 | 0.626 ||
| 30 | 301 | 0.812 |
| 40 | 401 | 1.253 |
| 50 | 501 |  1.489 |
| 100 | 101 | 3.166 |
| 150 | 1501 | 4.403 |
| 200 | 2001 | 6.384 |

at a size of ~ 1500 the cache becomes detrimental for this use case and degrades performance.

### After

| Years  | Cache Size | Time (seconds) |
| ------ | ----------- | ---------------- |
| 1 | 11 | 0.041 |
| 5 | 51 | 0.035 |
| 10 | 101 | 0.031 |
| 15 | 151 | 0.032 |
| 20 | 201 | 0.032 |
| 30 | 301 | 0.035 |
| 40 | 401 | 0.032 |
| 50 | 501 | 0.032 |
| 100 | 1001 | 0.045 |
| 150 | 1501 | 0.042 |
| 200 | 2001 | 0.041 |