@ptrimble 2017-02-24
Wow, thank you for this! And thank you for your patience, I am returning from paternity leave so it's taking me a while to get back up to running speed.

The changes initially look great! I would like to finish up some in-progress work I left from before I took my time off and then this is next on my list. I'll be in touch soon!